% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnscc.R
\docType{data}
\name{hnscc}
\alias{hnscc}
\title{High dimensional head and neck cancer survival and gene expression data}
\format{
A dataframe with 565 rows and 104 variables
\describe{
\item{ID}{"Column/Variable name" consisting id of subjects}
\item{Death}{"Column/Variable name" consisting survival event}
\item{OS}{"Column/Variable name" consisting duration of overall survival}
\item{PFS}{"Column/Variable name" consisting duration of progression free survival}
\item{Prog}{"Column/Variable name" consisting progression event}
\item{GJB1,...,HMGCS2}{High dimensional covariates}}
}
\usage{
hnscc
}
\description{
High dimensional head and neck cancer gene expression data
}
\examples{
data(hnscc)
}
\keyword{datasets}
