% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srdata.R
\docType{data}
\name{srdata}
\alias{srdata}
\title{High dimensional protein gene expression data}
\format{
A dataframe with 288 rows and 250 variables
\describe{
\item{ID}{"Column/Variable name" consisting id of subjects}
\item{Visit}{"Column/Variable name" consisting number of times observations recorded}
\item{event}{"Column/Variable name" consisting survival event}
\item{OS}{"Column/Variable name" consisting duration of overall survival}
\item{C6kine,.....,GFRalpha4}{High dimensional covariates}}
}
\usage{
srdata
}
\description{
High dimensional protein gene expression data
}
\examples{
data(srdata)
}
\keyword{datasets}
